Crate fingerprint_core

fingerprint-core

System-level protection core abstract layer

From single service protection improve to system-level protection, provides system-level core abstractions and interface.

Core Positioning

fingerprint-core is system-level protection core, all external components revolve around this core:

System-level abstractions: system context, network traffic, protection decision etc.

Offense and defense unified interface: fingerprint abstractions, analysis interface, protection interface etc.

Core types and utilities: type definitions, metadata, utility functions etc.

Core Features

System-level Abstractions

System context (SystemContext): complete network entity information (IP, port, protocol, direction etc.)

Network traffic (NetworkFlow): system level network traffic, including context and fingerprint info

System protection interface (SystemProtector): system-level protection unified interface

System analysis interface (SystemAnalyzer): unified system-level analysis interface

Offense and defense unified abstractions

fingerprint abstractions (Fingerprint trait): support TLS, HTTP, TCP etc.multiple fingerprint types

fingerprintmetadata (FingerprintMetadata): including browser, operating system, confidence etc.info

TLS fingerprint (ClientHelloSignature): TLS ClientHello signature

HTTP fingerprint (HttpFingerprint): HTTP request fingerprint

TCP fingerprint (TcpFingerprint): TCP connection fingerprint

core types and utilities

type system: BrowserType, OperatingSystem etc.coretype

utility functions: GREASE process, randomly select etc.utility functions

Re-exports

pub use error::ConfigError;

pub use error::DatabaseError;

pub use error::DnsError;

pub use error::FingerprintError;

pub use error::HttpError;

pub use error::ParseError;

pub use error::RateLimitError;

pub use error::Result;

pub use error::TcpError;

pub use error::TlsError;

pub use error::ValidationError;

pub use cache::CacheError;

pub use metrics::record_cache_hit;

pub use metrics::record_cache_miss;

pub use metrics::record_db_operation;

pub use metrics::record_error;

pub use metrics::record_fingerprint_duration;

pub use metrics::record_ml_inference;

pub use metrics::ANOMALY_DETECTION_TOTAL;

pub use metrics::ANOMALY_FALSE_POSITIVE_RATE;

pub use metrics::ANOMALY_SCORE;

pub use metrics::CACHE_EVICTIONS_TOTAL;

pub use metrics::CACHE_HIT_RATE;

pub use metrics::CACHE_MISS_RATE;

pub use metrics::CACHE_SIZE_BYTES;

pub use metrics::CPU_USAGE_PERCENT;

pub use metrics::DB_CONNECTIONS_ACTIVE;

pub use metrics::DB_OPERATION_DURATION_MS;

pub use metrics::DB_QUERIES_TOTAL;

pub use metrics::DNS_CACHE_HIT_RATE;

pub use metrics::DNS_RESOLUTION_DURATION_MS;

pub use metrics::DNS_RESOLUTION_TOTAL;

pub use metrics::ERRORS_TOTAL;

pub use metrics::ERROR_RATE;

pub use metrics::FINGERPRINT_RECOGNITION_DURATION_MS;

pub use metrics::FINGERPRINT_RECOGNITION_TOTAL;

pub use metrics::FINGERPRINT_SIMILARITY_SCORE;

pub use metrics::GOROUTINES_ACTIVE;

pub use metrics::HTTP_POOL_CONNECTIONS;

pub use metrics::HTTP_REQUEST_DURATION_MS;

pub use metrics::HTTP_REQUEST_TOTAL;

pub use metrics::JA_FINGERPRINT_CALC_TOTAL;

pub use metrics::MEMORY_USAGE_MB;

pub use metrics::ML_INFERENCE_DURATION_MS;

pub use metrics::ML_PREDICTION_ACCURACY;

pub use metrics::ML_PREDICTION_TOTAL;

pub use metrics::RATE_LIMIT_CHECK_TOTAL;

pub use metrics::RATE_LIMIT_QUOTA_USAGE;

pub use metrics::RATE_LIMIT_REJECTIONS_TOTAL;

pub use metrics::TLS_CLIENTHELLO_PARSE_MS;

pub use metrics::TLS_FINGERPRINT_GENERATION_TOTAL;

pub use fingerprint::Fingerprint;

pub use fingerprint::FingerprintComparator;

pub use fingerprint::FingerprintComparison;

pub use fingerprint::FingerprintType;

pub use metadata::FingerprintMetadata;

pub use grease::filter_grease_values;

pub use grease::get_random_grease;

pub use grease::is_grease_value;

pub use grease::remove_grease_values;

pub use grease::TLS_GREASE_VALUES;

pub use hassh::HASSHServer;

pub use hassh::SSHKexInit;

pub use hassh::HASSH;

pub use hassh::JA4SSH;

pub use ja3::JA3;

pub use ja3::JA3S;

pub use ja4::ConsistencyReport;

pub use ja4::JA4;

pub use ja4::JA4H;

pub use ja4::JA4L;

pub use ja4::JA4S;

pub use ja4::JA4T;

pub use ja4::JA4X;

pub use signature::ClientHelloSignature;

pub use version::TlsVersion;

pub use pqc::PQCAlgorithm;

pub use pqc::PQCBrowserSupport;

pub use pqc::PQCCapabilities;

pub use wasm::WasmBrowserSupport;

pub use wasm::WasmCapabilities;

pub use wasm::WasmMemoryFingerprint;

pub use wasm::WasmTableFingerprint;

pub use wasm::WasmVersion;

pub use hpack::static_table;

pub use hpack::DynamicTableEntry;

pub use hpack::DynamicTableSnapshot;

pub use hpack::EncodedHeaderField;

pub use hpack::HpackAnalyzer;

pub use hpack::HpackFingerprint;

pub use hpack::HpackHeaderList;

pub use hpack::HuffmanEncoding;

pub use hpack::IndexType;

pub use hpack::StaticTableEntry;

pub use http::Http2Settings;

pub use http::HttpFingerprint;

pub use http2_frame_parser::find_settings_frame;

pub use http2_frame_parser::is_http2_connection;

pub use http2_frame_parser::Http2FrameHeader;

pub use http2_frame_parser::Http2FrameType;

pub use http2_frame_parser::Http2ParseError;

pub use http2_frame_parser::Http2SettingsFrame;

pub use http2_frame_parser::Http2SettingsMatcher;

pub use http2_frame_parser::HTTP2_PREFACE;

pub use packet_capture::EthernetHeader;

pub use packet_capture::Ipv4Header;

pub use packet_capture::Ipv6Header;

pub use packet_capture::NetworkProtocol;

pub use packet_capture::PacketFlowAnalyzer;

pub use packet_capture::PacketParser;

pub use packet_capture::ParsedPacket;

pub use packet_capture::PcapGlobalHeader;

pub use packet_capture::PcapPacketHeader;

pub use packet_capture::TcpFlow;

pub use packet_capture::TcpHeader;

pub use packet_capture::TransportProtocol;

pub use packet_capture::UdpHeader;

pub use tcp::TcpFingerprint;

pub use tcp::TcpProfile;

pub use tcp_handshake::signatures;

pub use tcp_handshake::AckCharacteristics;

pub use tcp_handshake::IpCharacteristics;

pub use tcp_handshake::SynAckCharacteristics;

pub use tcp_handshake::SynCharacteristics;

pub use tcp_handshake::TcpFlags;

pub use tcp_handshake::TcpHandshakeAnalyzer;

pub use tcp_handshake::TcpHandshakeFingerprint;

pub use tcp_handshake::TcpOption;

pub use tcp_handshake::TcpOptionType;

pub use types::BrowserType;

pub use types::OperatingSystem;

pub use types::OperatingSystems;

pub use types::UserAgentTemplate;

pub use types::OPERATING_SYSTEMS;

pub use utils::extract_chrome_version;

pub use utils::extract_platform;

pub use utils::infer_browser_from_profile_name;

pub use utils::is_mobile_profile;

pub use utils::random_choice;

pub use utils::random_choice_string;

pub use benchmark::Benchmark;

pub use benchmark::CacheBenchmark;

pub use benchmark::CacheBenchmarkSuite;

pub use benchmark::HttpMetrics;

pub use benchmark::Timer;

pub use rate_limiting::current_unix_timestamp;

pub use rate_limiting::EndpointConfig;

pub use rate_limiting::MetricsSnapshot;

pub use rate_limiting::QuotaTier;

pub use rate_limiting::RateLimitResponse;

pub use rate_limiting::RateLimiter;

pub use rate_limiting::UserQuota;

pub use rate_limiting_redis::RedisBackendError;

pub use rate_limiting_redis::RedisConfig;

pub use rate_limiting_redis::RedisQuotaEntry;

pub use rate_limiting_redis::RedisRateLimitBackend;

pub use rate_limiting_redis::RedisResult;

pub use rate_limiting_metrics::MetricsHandler;

pub use rate_limiting_metrics::PrometheusMetrics;

pub use rate_limiting_metrics::TierMetrics;

pub use cache::Cache;

pub use cache::CacheResult;

pub use cache::CacheStats;

pub use cache::CacheTTL;

pub use cache::CacheTier;

pub use cache::DistributedLock;

pub use cache::LockGuard;

pub use cache_redis::RedisCacheConfig;

pub use system::AnalysisDetails;

pub use system::FlowCharacteristics;

pub use system::NetworkFlow;

pub use system::ProtocolType;

pub use system::SystemAnalysisResult;

pub use system::SystemAnalyzer;

pub use system::SystemContext;

pub use system::SystemProtectionDecision;

pub use system::SystemProtectionResult;

pub use system::SystemProtectionStats;

pub use system::SystemProtector;

pub use system::ThreatType;

pub use system::TrafficDirection;

pub use dicttls::*;

Modules

benchmark

Performance benchmarking utilities for fingerprint-rust

cache

cache_redis

database

Fingerprint Database

dicttls

TLS dictionary module

error

Comprehensive error types for fingerprint operations

fingerprint

Fingerprint core abstraction

grease

GREASE valueprocessmodule

hassh

HASSH SSH fingerprintimplement

hpack

http

HTTP fingerprintcoretype

http2_frame_parser

HTTP/2 Frame Parser

ja3

JA3/JA3S TLS fingerprintimplement

ja4

JA4+ fingerprintseriesimplement

ja3_database

JA3 Fingerprint Database

jarm

JARM: Active TLS Server Fingerprinting

metadata

fingerprintmetadata

metrics

Comprehensive metrics collection for fingerprint-rust

packet_capture

pcap_generator

pqc

Post-Quantum Cryptography (PQC) Detection

rate_limiting

rate_limiting_metrics

rate_limiting_redis

signature

TLS ClientHello Signature module

system

system-level protectioncoreabstract

tcp

TCP Fingerprint Type

tcp_handshake

tls_parser

TLS ClientHello Parser

types

type definitionsmodule

utils

utility functionsmodule

version

TLS versionmodule

wasm

WebAssembly (WASM) Fingerprinting Detection